use same transform to put elements in the middle of page
