+++ server/lib/transforms/insert-ad.js
@@ -1,9 +1,9 @@
 'use strict';
 
+const getMiddleParagraph = require('./get-middle-paragraph');
 const adMarkup = () => `<div class="ad-container"><amp-ad width="300" height="250" type="doubleclick"></amp-ad></div>`;
 
 module.exports = function insertAd($, options) {
-		const paragraph = $('p').eq(3);
-		paragraph.after(adMarkup(options.targeting));
+	getMiddleParagraph($).after(adMarkup(options.targeting));
 	return $;
 };
