+++ server/lib/transforms/light-signup.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const url = require('url');
+const getMiddleParagraph = require('./get-middle-paragraph');
 
 const formatLightSignupUrl = options => url.format(Object.assign(
 	url.parse(options.lightSignupUrl),
@@ -26,30 +27,7 @@ const lightSignupMarkup = options => `<div amp-access="NOT session" amp-access-h
 
 module.exports = function addLightSignup($, options) {
 	if(options.enableLightSignup) {
-		const paras = $.root().children('p');
-
-		// Zero-indexed position
-		const idealPosition = Math.max(3, Math.floor(paras.length / 2)) - 1;
-		const maxPosition = paras.length - 1;
-
-		let position = idealPosition;
-
-		// Try to position in the middle of the article, working up until a suitable
-		// place can be found
-		while(!paras.eq(position).next().is('p') && position >= 3) {
-			position--;
-		}
-
-		// If no place can be found in the first half of the article, place at some point
-		// after the middle
-		if(position < 3) {
-			position = idealPosition;
-			while(!paras.eq(position).next().is('p') && position < maxPosition) {
-				position++;
-			}
-		}
-
-		paras.eq(position).after(lightSignupMarkup(options));
+		getMiddleParagraph($).after(lightSignupMarkup(options));
 	}
 
 	return $;
